move upload functions out of testing
