+++ modules/renter/upload.go
@@ -2,25 +2,107 @@ package renter
 
 import (
 	"errors"
+	"io"
 	"os"
 	"path/filepath"
-	"time"
 
 	"github.com/NebulousLabs/Sia/crypto"
 	"github.com/NebulousLabs/Sia/modules"
 	"github.com/NebulousLabs/Sia/types"
 )
 
-const (
-	maxUploadAttempts = 3
-	parallelUploads   = 3
-)
+type uploadPiece struct {
+	data       []byte
+	chunkIndex uint64
+	pieceIndex uint64
+}
 
-var (
-	errUploadFailed = errors.New("failed to upload to the desired host")
+// An uploader uploads pieces to a host. This interface exists to facilitate
+// easy testing.
+// TODO: once fileContracts are stored in the renter, have these return IDs?
+type uploader interface {
+	// connect initiates the connection to the uploader.
+	connect() (fileContract, error)
 
-	redundancy = 8
-)
+	// addPiece uploads a piece to the uploader.
+	addPiece(uploadPiece) (fileContract, error)
+}
+
+// newFile creates a new file object.
+func newFile(ecc modules.ECC, pieceSize, fileSize uint64) *dfile {
+	key, _ := crypto.GenerateTwofishKey()
+	return &dfile{
+		Size:      fileSize,
+		Contracts: make(map[modules.NetAddress]fileContract),
+		MasterKey: key,
+		ecc:       ecc,
+		pieceSize: pieceSize,
+	}
+}
+
+// worker uploads pieces to a host as directed by reqChan. It sends the
+// updated fileContract down respChan.
+func (f *dfile) worker(host uploader, reqChan chan uploadPiece, respChan chan *fileContract) {
+	// TODO: move connect outside worker
+	_, err := host.connect()
+	if err != nil {
+		respChan <- nil
+		return
+	}
+	for req := range reqChan {
+		contract, err := host.addPiece(req)
+		if err != nil {
+			respChan <- nil
+			return // this host is now dead to us; upload will use a new one
+		}
+		respChan <- &contract
+	}
+}
+
+// upload reads chunks from r and uploads them to hosts. It spawns a worker
+// for each host, and instructs them to upload pieces of each chunk.
+func (f *dfile) upload(r io.Reader, hosts []uploader) error {
+	// create request/response channels and spawn workers
+	reqChans := make([]chan uploadPiece, len(hosts))
+	respChans := make([]chan *fileContract, len(hosts))
+	for i, h := range hosts {
+		reqChans[i] = make(chan uploadPiece)
+		respChans[i] = make(chan *fileContract)
+		go f.worker(h, reqChans[i], respChans[i])
+	}
+
+	chunk := make([]byte, f.chunkSize())
+	for i := uint64(0); ; i++ {
+		// read next chunk
+		n, err := io.ReadFull(r, chunk)
+		if err == io.EOF {
+			break
+		} else if err != nil && err != io.ErrUnexpectedEOF {
+			return err
+		}
+		// encode
+		pieces, err := f.ecc.Encode(chunk)
+		if err != nil {
+			return err
+		}
+		// send upload requests to workers
+		for j, ch := range reqChans {
+			ch <- uploadPiece{pieces[j], i, uint64(j)}
+		}
+		// read upload responses from workers
+		for _, ch := range respChans {
+			contract := <-ch
+			if contract == nil {
+				// choose new host somehow
+				//go f.worker(newhost, reqChans[j], respChans[j])
+				continue
+			}
+			f.Contracts[contract.IP] = *contract
+		}
+		f.uploaded += uint64(n)
+	}
+	return nil
+}
 
 // checkWalletBalance looks at an upload and determines if there is enough
 // money in the wallet to support such an upload. An error is returned if it is
@@ -34,7 +116,7 @@ func (r *Renter) checkWalletBalance(up modules.FileUploadParams) error {
 	curSize := types.NewCurrency64(uint64(fileInfo.Size()))
 
 	var averagePrice types.Currency
-	sampleSize := redundancy * 3 / 2
+	sampleSize := up.ECC.NumPieces() * 3 / 2
 	hosts := r.hostDB.RandomHosts(sampleSize)
 	for _, host := range hosts {
 		averagePrice = averagePrice.Add(host.Price)
@@ -53,38 +135,6 @@ func (r *Renter) checkWalletBalance(up modules.FileUploadParams) error {
 	return nil
 }
 
-// threadedUploadPiece will upload the piece of a file to a randomly chosen
-// host. If the wallet has insufficient balance to support uploading,
-// uploadPiece will give up. The file uploading can be continued using a repair
-// tool. Upon completion, the memory containg the piece's information is
-// updated.
-func (r *Renter) threadedUploadPiece(host modules.HostSettings, up modules.FileUploadParams, piece *filePiece) error {
-	// Set 'Repairing' for the piece to true.
-	lockID := r.mu.Lock()
-	piece.Repairing = true
-	r.mu.Unlock(lockID)
-
-	// Try 'maxUploadAttempts' hosts before giving up.
-	for attempts := 0; attempts < maxUploadAttempts; attempts++ {
-		// Negotiate the contract with the host. If the negotiation is
-		// unsuccessful, we need to try again with a new host.
-		err := r.negotiateContract(host, up, piece)
-		if err == nil {
-			return nil
-		}
-
-		// The previous attempt didn't work. We will try again after
-		// sleeping for a randomized amount of time to increase our chances
-		// of success. This will help spread things out if there are
-		// problems with network congestion or other randomized issues.
-		r, _ := crypto.RandIntn(256)
-		time.Sleep(100 * time.Millisecond * time.Duration(r))
-	}
-
-	// All attempts failed.
-	return errors.New("failed to upload filePiece")
-}
-
 // Upload takes an upload parameters, which contain a file to upload, and then
 // creates a redundant copy of the file on the Sia network.
 func (r *Renter) Upload(up modules.FileUploadParams) error {
@@ -94,7 +144,13 @@ func (r *Renter) Upload(up modules.FileUploadParams) error {
 		return errors.New("nickname and file name must have the same extension")
 	}
 
-	err := r.checkWalletBalance(up)
+	// Open the file.
+	handle, err := os.Open(up.Filename)
+	if err != nil {
+		return err
+	}
+
+	err = r.checkWalletBalance(up)
 	if err != nil {
 		return err
 	}
@@ -107,42 +163,28 @@ func (r *Renter) Upload(up modules.FileUploadParams) error {
 		return errors.New("file with that nickname already exists")
 	}
 
-	// Check that the file exists and is less than 500 MiB.
-	fileInfo, err := os.Stat(up.Filename)
+	// Check that the file is less than 5 GiB.
+	fileInfo, err := handle.Stat()
 	if err != nil {
 		return err
 	}
-	// NOTE: The upload max of 500 MiB is temporary and therefore does not have a
-	// constant. This should be removed once micropayments + upload resuming
-	// are in place. 500 MiB is chosen to prevent confusion - on anybody's
-	// machine any file appearing to be under 500 MB will be below the hard
+	// NOTE: The upload max of 5 GiB is temporary and therefore does not have
+	// a constant. This should be removed once micropayments + upload resuming
+	// are in place. 5 GiB is chosen to prevent confusion - on anybody's
+	// machine any file appearing to be under 5 GB will be below the hard
 	// limit.
-	if fileInfo.Size() > 500*1024*1024 {
-		return errors.New("cannot upload a file larger than 500 MB")
+	if fileInfo.Size() > 5*1024*1024*1024 {
+		return errors.New("cannot upload a file larger than 5 GB")
 	}
 
-	// Check that the hostdb is sufficiently large to support an upload. Right
-	// now that value is set to 3, but in the future the logic will be a bit
-	// more complex; once there is erasure coding we'll want to hit the minimum
-	// number of pieces plus some buffer before we decide that an upload is
-	// okay.
-	if len(r.hostDB.ActiveHosts()) < 1 {
+	// Check that the hostdb is sufficiently large to support an upload.
+	// TODO: ActiveHosts needs to only report hosts >= v0.4
+	if len(r.hostDB.ActiveHosts()) < up.ECC.NumPieces() {
 		return errors.New("not enough hosts on the network to upload a file")
 	}
 
 	// Create file object.
-	f := &file{
-		Name: up.Nickname,
-
-		PiecesRequired: 1,
-		Pieces:         make([]filePiece, up.Pieces),
-		UploadParams:   up,
-		renter:         r,
-	}
-	for i := range f.Pieces {
-		f.Pieces[i].Repairing = true
-		f.Pieces[i].PieceSize = uint64(fileInfo.Size())
-	}
+	f := newFile(up.ECC, up.PieceSize, uint64(fileInfo.Size()))
 
 	// Add file to renter.
 	lockID = r.mu.Lock()
@@ -150,51 +192,21 @@ func (r *Renter) Upload(up modules.FileUploadParams) error {
 	r.save()
 	r.mu.Unlock(lockID)
 
-	// Upload to hosts in parallel. To facilitate this, we create channels of
-	// hosts and file pieces, and spawn goroutines that attempt to match each
-	// piece to a host.
-	hostPool := make(chan modules.HostSettings, 3*redundancy)
-	for _, host := range r.hostDB.RandomHosts(3 * redundancy) {
-		hostPool <- host
+	// Upload to hosts in parallel.
+	var hosts []uploader
+	// TODO: hostDB needs to provide uploaders, or we need to convert them somehow
+	for range r.hostDB.RandomHosts(up.ECC.NumPieces()) {
+		hosts = append(hosts, nil)
 	}
-	piecePool := make(chan *filePiece, len(f.Pieces))
-	for i := range f.Pieces {
-		piecePool <- &f.Pieces[i]
-	}
-	errChan := make(chan error, len(f.Pieces))
-	for i := 0; i < parallelUploads; i++ {
-		go func() {
-			for piece := range piecePool {
-				var err error
-				for host := range hostPool {
-					err = r.threadedUploadPiece(host, up, piece)
-					if err == nil {
-						break
-					}
-				}
-				errChan <- err
-			}
-		}()
+	err = f.upload(handle, hosts)
+	if err != nil {
+		// Upload failed; remove the file object.
+		lockID = r.mu.Lock()
+		delete(r.files, up.Nickname)
+		r.save()
+		r.mu.Unlock(lockID)
+		return errors.New("failed to upload any file pieces")
 	}
 
-	// Wait for success or failure. Since we are (currently) using full
-	// replication, success means "one piece was uploaded," while failure
-	// means "zero pieces were uploaded."
-	reqPieces := f.PiecesRequired
-	for i := 0; i < up.Pieces; i++ {
-		if <-errChan == nil {
-			reqPieces--
-			if reqPieces <= 0 {
-				return nil
-			}
-		}
-	}
-
-	// All uploads failed. Remove the file object.
-	lockID = r.mu.Lock()
-	delete(r.files, up.Nickname)
-	r.save()
-	r.mu.Unlock(lockID)
-
-	return errors.New("failed to upload any file pieces")
+	return nil
 }
